+++ enterprise/e2e/auth-sso/hurl/claims.all.hurl
@@ -139,7 +139,7 @@ jsonpath "$.entries[1].type" == "directory"
 jsonpath "$.entries[1].path" == "/self/"
 jsonpath "$.entries[1].private" == false
 jsonpath "$.entries[1].schemas" == 65
-jsonpath "$.entries[1].health" == 100
+jsonpath "$.entries[1].health" == 99
 
 POST {{base}}/self/v1/api/schemas/evaluate{{root_listing_schema}}
 ```
