+++ enterprise/e2e/auth/hurl/directory.all.hurl
@@ -39,7 +39,7 @@ jsonpath "$.entries[2].type" == "directory"
 jsonpath "$.entries[2].path" == "/self/"
 jsonpath "$.entries[2].private" == false
 jsonpath "$.entries[2].schemas" == 65
-jsonpath "$.entries[2].health" == 100
+jsonpath "$.entries[2].health" == 99
 POST {{base}}/self/v1/api/schemas/evaluate{{root_listing_schema}}
 ```
 {{root_listing}}
@@ -104,7 +104,7 @@ jsonpath "$.entries[4].type" == "directory"
 jsonpath "$.entries[4].path" == "/self/"
 jsonpath "$.entries[4].private" == false
 jsonpath "$.entries[4].schemas" == 65
-jsonpath "$.entries[4].health" == 100
+jsonpath "$.entries[4].health" == 99
 
 POST {{base}}/self/v1/api/schemas/evaluate{{reports_listing_schema}}
 Authorization: Bearer primary-secret-key
