+++ enterprise/e2e/html/hurl/list.all.hurl
@@ -33,7 +33,7 @@ jsonpath "$.entries[1].private" == false
 jsonpath "$.entries[1].title" == "Self"
 jsonpath "$.entries[1].description" == "The schemas that define the current version of this instance"
 jsonpath "$.entries[1].type" == "directory"
-jsonpath "$.entries[1].health" == 100
+jsonpath "$.entries[1].health" == 99
 jsonpath "$.entries[1].schemas" == 65
 jsonpath "$.entries[1].path" == "/self/"
 jsonpath "$.entries[2].name" == "test"
