+++ src/router/CMakeLists.txt
@@ -11,6 +11,7 @@ target_link_libraries(sourcemeta_one_router PUBLIC sourcemeta::core::jsonrpc)
 target_link_libraries(sourcemeta_one_router PUBLIC sourcemeta::one::metapack)
 target_link_libraries(sourcemeta_one_router PUBLIC sourcemeta::blaze::evaluator)
 target_link_libraries(sourcemeta_one_router PUBLIC sourcemeta::blaze::compiler)
+target_link_libraries(sourcemeta_one_router PUBLIC sourcemeta::blaze::bundle)
 target_link_libraries(sourcemeta_one_router PUBLIC sourcemeta::blaze::output)
 target_link_libraries(sourcemeta_one_router PRIVATE sourcemeta::core::http)
 target_link_libraries(sourcemeta_one_router PRIVATE sourcemeta::core::uri)
