+++ test/e2e/empty/hurl/test.all.hurl
@@ -37,5 +37,5 @@ jsonpath "$.entries[0].type" == "directory"
 jsonpath "$.entries[0].title" == "Self"
 jsonpath "$.entries[0].description" == "The schemas that define the current version of this instance"
 jsonpath "$.entries[0].path" == "/self/"
-jsonpath "$.entries[0].health" == 100
+jsonpath "$.entries[0].health" == 99
 jsonpath "$.entries[0].schemas" == 65
