+++ test/e2e/html/hurl/list.all.hurl
@@ -27,7 +27,7 @@ jsonpath "$.entries[0].private" == false
 jsonpath "$.entries[0].title" == "Self"
 jsonpath "$.entries[0].description" == "The schemas that define the current version of this instance"
 jsonpath "$.entries[0].type" == "directory"
-jsonpath "$.entries[0].health" == 100
+jsonpath "$.entries[0].health" == 99
 jsonpath "$.entries[0].schemas" == 65
 jsonpath "$.entries[0].path" == "/self/"
 jsonpath "$.entries[1].name" == "test"
